Rework Library screens by seperating them out to their own classes.
